new responsive vertical pane working
