+++ src/tools/style-tools.ts
@@ -5,11 +5,12 @@ export function makeClassName(classNames: string[]): string {
 export const colors = {
     $bg: '#1b2426',
     $bgNav: '#2e3d51',
+    $blue: '#5ba1cf',
     $scrollbarThumb:'#2e4b52',
     $silver100: '#fff',
     $silver200: '#b9bdb6',
     $silver300: '#8c8e8a',
-    $blue: '#5ba1cf',
+    $purple: '#ce61fb',
     $react: '#61dafb',
     $red: '#e24a4a',
     $redConsole: '#3c0303',
@@ -21,6 +22,8 @@ export const colors = {
 // FB8161 => coomplementare react
 
 export const generalBorderStyle = `1px solid ${colors.$silver300}`
+
+export const responsiveBreakpoint = 540;
 
 export const transitionDuration = {
     fast: '100ms',
